config/requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dbfread==2.0.7
Django==4.2.7
django-apscheduler==0.6.2
django-cors-headers==4.3.1
django-crontab==0.7.1
django-dotenv==1.4.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==21.2.0
idna==3.4
kombu==5.3.4
mysqlclient==2.2.0
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
prompt-toolkit==3.0.41
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
